Make xCharts show up some dummy data
Unfortunately xCharts isn't that compatible with browserify so had to
skip using it. Still need to clue pingdom data with xCharts.

Likely the nicest solution is to output the data to some JSON which will
be then read by the service. As a side benefit we get something that is
easy to display elsewhere. A good thing at least for now.