Use pingdom summary API to fetch latency and uptime data
This is very effective request-wise. It costs around 1 + <amount of
checks> to get the latency and uptime info.

Currently this is performed once per day using cron although that can be
changed if needed.

Note that the frontend still needs to be updated to use the new data
format. It is more compact than the previous one. There still could be
some problems in it, though.